moved utils into its own folder and renamed. minor improvements to the readme
