+++ frontend/src/adapters/user-adapter.js
@@ -1,6 +1,6 @@
 // These functions all take in a body and return an options object
 // with the provided body and the remaining options
-import { fetchHandler, getPostOptions, getPatchOptions } from "../utils";
+import { fetchHandler, getPostOptions, getPatchOptions } from "../utils/fetchingUtils";
 
 const baseUrl = '/api/users';
